refactor(InlineCssVariablesInSvg): Updated according to #174 reviews
