+++ package.json
@@ -5,7 +5,6 @@
   "main": "index.js",
   "dependencies": {
     "@mobily/ts-belt": "^3.12.0",
-    "@types/color": "^3.0.3",
     "@types/lodash": "^4.14.175",
     "@types/mustache": "^4.1.2",
     "@types/node": "16.3.3",
@@ -14,7 +13,6 @@
     "@types/theme-ui": "^0.6.0",
     "@types/tinycolor2": "^1.4.3",
     "@types/xml2js": "^0.4.9",
-    "color": "^4.2.3",
     "deepmerge": "^4.2.2",
     "got": "^11.8.5",
     "lodash": "^4.17.21",
